P1: Optimization progress (C89)
Further progress on implementing the arithmetic sequence solution for an arbitrary number of divisors. Sequence solution will count some values multiple times when they are divisible by more than one given divisor, requiring a table of possible products to be generated in order to adequately correct the end sum. Current product table function should generate all possible products of the first value of the divisor set multiplied by any other value or subset of values, but will not do the same for subsequent values in the set. A new loop is likely required to make this possible.